Integrate all planar pcs tests into a single file
